feat(src): extract WebSocket classes to src/websocket/ module
Closes #209
