+++ MistHelper.py
@@ -55,8 +55,9 @@ if TYPE_CHECKING:
     import pyte
     import requests
     import urllib3
-    import websocket
     from prettytable import PrettyTable
+
+    import websocket
 
 # ============================================================================
 # POLYGLOT DATABASE LAYER (OPTIONAL)
@@ -75,6 +76,10 @@ from src.org_data_collector import OrgDataCollector
 from src.ssh.ssh_runner import EnhancedSSHRunner
 from src.wan_hub_group_manager import WanHubGroupNumberManager
 from src.wan_vpn_builder import WanVpnBuilder
+from src.websocket.commands import WebSocketCommands
+from src.websocket.context import WebSocketCmdDeps
+from src.websocket.diag_commands import WebSocketNetworkDiagCommands
+from src.websocket.manager import WebSocketManager
 
 # ============================================================================
 # EARLY LOGGING SETUP
@@ -5726,853 +5731,6 @@ class DeviceDataFetcher:
         processed = DataProcessingUtils.escape_multiline(processed)  # type: ignore[no-untyped-call]
         DataExporter.save_data_to_output(processed, self.filename)  # type: ignore[no-untyped-call]
         DisplayUtils.dict_list_as_pretty_table(processed)
-
-
-# ============================================================================
-# WEBSOCKET MANAGEMENT CLASS
-# ============================================================================
-
-
-class WebSocketManager:
-    """
-    WebSocket Manager for Mist API real-time communications.
-
-    This class handles WebSocket connections to the Mist API following the
-    documented patterns: subscribe first, issue POST command, await results.
-
-    SECURITY: WebSocket connections use authenticated sessions with proper
-    credential handling and session-based command demultiplexing.
-    """
-
-    def __init__(self, mist_session: Any, mist_host: str | None = None) -> None:
-        """
-        Initialize WebSocket manager with Mist session.
-
-        Args:
-            mist_session: Authenticated Mist API session
-            mist_host: Mist API host (if None, will get from session)
-        """
-        self.mist_session = mist_session
-        self.mist_host = mist_host or getattr(mist_session, "host", None) or os.getenv("MIST_HOST", "api.mist.com")
-
-        # Convert API host to WebSocket host
-        assert self.mist_host is not None, "mist_host must be set"  # nosec B101
-        websocket_host = self.mist_host.replace("api.", "api-ws.")
-        self.websocket_url = f"wss://{websocket_host}/api-ws/v1/stream"
-        self.websocket_connection: websocket.WebSocketApp | None = None
-        self.logger = logging.getLogger(__name__)
-        self.connected = False
-        self.subscribed_channels: set[str] = set()
-        self.confirmed_subscriptions: set[str] = set()
-
-        # Results storage for command outputs
-        self.command_results: dict[str, Any] = {}
-        self.results_lock = threading.Lock()
-
-    def connect(self) -> bool:
-        """
-        Establish WebSocket connection with proper authentication.
-
-        Returns:
-            bool: True if connection successful, False otherwise
-        """
-        try:
-            # Prepare authentication headers using session token
-            mist_apitoken = getattr(self.mist_session, "apitoken", None) or os.getenv("MIST_APITOKEN")
-            if not mist_apitoken:
-                self.logger.error("No API token found in session or environment")
-                return False
-
-            self.logger.debug(f"WebSocket URL: {self.websocket_url}")
-            self.logger.debug(f"Auth token configured (length: {len(mist_apitoken)} chars)")
-            auth_header = f"Authorization: Token {mist_apitoken}"
-            headers = [auth_header]
-
-            # Create WebSocket connection
-            self.websocket_connection = websocket.WebSocketApp(
-                self.websocket_url,
-                header=headers,
-                on_message=self._on_message,
-                on_error=self._on_error,
-                on_close=self._on_close,
-                on_open=self._on_open,
-            )
-
-            # Start connection in background thread
-            connection = self.websocket_connection
-            self.websocket_thread = threading.Thread(target=connection.run_forever, daemon=True)
-            self.websocket_thread.start()
-
-            # Wait for connection to establish
-            timeout_counter = 0
-            while not self.connected and timeout_counter < 10:
-                time.sleep(0.5)
-                timeout_counter += 1
-                if timeout_counter % 2 == 0:
-                    self.logger.debug(f"WebSocket handshake waiting... ({timeout_counter * 0.5:.1f}s)")
-
-            if self.connected:
-                self.logger.info("WebSocket connection established successfully")
-                return True
-            else:
-                self.logger.error("WebSocket connection timeout")
-                return False
-
-        except Exception as connection_error:
-            self.logger.error(f"WebSocket connection failed: {connection_error}")
-            return False
-
-    def subscribe_to_channel(self, channel_path: str) -> bool:
-        """
-        Subscribe to a WebSocket channel for receiving command outputs.
-
-        Args:
-            channel_path (str): Channel path (e.g., "/sites/{site_id}/devices/{device_id}/cmd")
-
-        Returns:
-            bool: True if subscription successful, False otherwise
-        """
-        if not self.connected:
-            self.logger.error("Cannot subscribe: WebSocket not connected")
-            return False
-
-        try:
-            subscription_message = {"subscribe": channel_path}
-
-            if self.websocket_connection is not None:
-                self.websocket_connection.send(json.dumps(subscription_message))
-            self.subscribed_channels.add(channel_path)
-            self.logger.debug(f"Subscribed to channel: {channel_path}")
-            return True
-
-        except Exception as subscription_error:
-            self.logger.error(f"Channel subscription failed: {subscription_error}")
-            return False
-
-    def wait_for_subscription_confirmation(self, channel_path, timeout_seconds=10):  # type: ignore[no-untyped-def]
-        """
-        Wait for WebSocket subscription confirmation for a specific channel.
-
-        Args:
-            channel_path (str): Channel path to wait for confirmation
-            timeout_seconds (int): Maximum time to wait for confirmation
-
-        Returns:
-            bool: True if confirmation received, False if timeout
-        """
-        import time
-
-        start_time = time.time()
-
-        debug_mode = getattr(self, "debug_mode", False) or os.getenv("DEBUG", "").lower() in ["true", "1", "yes"]
-
-        if debug_mode:
-            self.logger.debug(f"Waiting for subscription confirmation for: {channel_path}")
-            print(f"[DEBUG] Waiting for subscription confirmation for: {channel_path}")
-
-        while time.time() - start_time < timeout_seconds:
-            # Check if subscription is confirmed
-            if channel_path in self.confirmed_subscriptions:
-                if debug_mode:
-                    self.logger.debug(f"Subscription confirmed for: {channel_path}")
-                    print(f"[DEBUG] Subscription confirmed for: {channel_path}")
-                return True
-
-            time.sleep(0.1)  # Small sleep to avoid busy waiting
-
-        # Timeout reached
-        if debug_mode:
-            self.logger.debug(f"Timeout waiting for subscription confirmation: {channel_path}")
-            print(f"[DEBUG] Timeout waiting for subscription confirmation: {channel_path}")
-        self.logger.warning(f"Timeout waiting for subscription confirmation: {channel_path}")
-        return False
-
-    def wait_for_command_result(  # noqa: C901, PLR0912, PLR0915
-        self,
-        session_id: str,
-        timeout_seconds: int = 30,
-        activity_timeout_seconds: int | None = None,
-    ) -> dict[str, Any] | None:
-        """
-        Wait for command result with specific session ID.
-
-        For commands like ping that produce multiple output segments,
-        this will collect all results until the command completes.
-
-        Args:
-            session_id (str): Session ID from command POST response
-            timeout_seconds (int): Maximum time to wait for result
-            activity_timeout_seconds (int): Seconds to wait after last message before considering complete
-
-        Returns:
-            dict: Complete command result data or None if timeout
-        """
-        import time
-
-        debug_mode = is_debug_mode()  # type: ignore[no-untyped-call]
-        start_time = time.time()
-        last_activity = time.time()
-        last_message_count = 0  # Track number of messages to detect new activity
-
-        # Reset MAC table completion cache for new sessions
-        if hasattr(self, "_mac_expected_entries"):
-            delattr(self, "_mac_expected_entries")
-
-        # Use custom activity timeout if provided, otherwise default to 2 seconds
-        activity_timeout = activity_timeout_seconds if activity_timeout_seconds is not None else 2
-        check_count = 0
-        last_debug_time = start_time
-        performance_log_interval = 5.0  # Log performance every 5 seconds
-
-        # Create performance monitor to detect infinite loops
-        perf_monitor = PerformanceMonitor(  # type: ignore[no-untyped-call]
-            f"wait_for_command_result({session_id[:8]}...)", max_iterations=10000, log_interval=5.0
-        )
-
-        if debug_mode:
-            self.logger.debug(f"Waiting for session {session_id} (timeout: {timeout_seconds}s)")
-            self.logger.debug(f"Current time: {time.time()}")
-            self.logger.debug(f"Activity timeout: {activity_timeout}s)")
-            print(f"[DEBUG] Waiting for session {session_id} (timeout: {timeout_seconds}s)")
-            print(f"[DEBUG] Current time: {time.time()}")
-            print(f"[DEBUG] Activity timeout: {activity_timeout}s)")
-
-        while time.time() - start_time < timeout_seconds:
-            # Monitor for infinite loops
-            perf_monitor.check_iteration()  # type: ignore[no-untyped-call]
-
-            current_time = time.time()
-            check_count += 1
-
-            # Performance logging every 5 seconds in debug mode
-            if debug_mode and (current_time - last_debug_time) >= performance_log_interval:
-                elapsed = current_time - start_time
-                self.logger.debug(f"Check #{check_count} at {elapsed:.1f}s - Still waiting for session {session_id}")
-                self.logger.debug(f"Last activity: {current_time - last_activity:.1f}s ago")
-                print(f"[PERF] Check #{check_count} at {elapsed:.1f}s - Still waiting for session {session_id}")
-                print(f"[PERF] Last activity: {current_time - last_activity:.1f}s ago")
-                with self.results_lock:
-                    available_sessions = list(self.command_results.keys())
-                    if session_id in self.command_results:
-                        msg_count = len(self.command_results[session_id])
-                        self.logger.debug(f"Found {msg_count} messages for our session")
-                        print(f"[PERF] Found {msg_count} messages for our session")
-                    else:
-                        self.logger.debug(f"Our session not in results yet. Available: {available_sessions}")
-                        print(f"[PERF] Our session not in results yet. Available: {available_sessions}")
-                last_debug_time = current_time
-
-            with self.results_lock:
-                if session_id in self.command_results:
-                    collected_output = self.command_results[session_id]
-                    current_message_count = len(collected_output)
-
-                    # Update last_activity only when NEW messages arrive
-                    if current_message_count > last_message_count:
-                        last_activity = time.time()
-                        last_message_count = current_message_count
-                        if debug_mode:
-                            self.logger.debug(
-                                f"New activity detected: {current_message_count} messages (+{current_message_count - (last_message_count - (current_message_count - last_message_count))}) "  # noqa: E501
-                            )
-
-                    if collected_output:
-                        # Check ALL messages for completion indicators, not just the latest
-                        all_raw_content = ""
-                        for result in collected_output:
-                            all_raw_content += result.get("raw", "")
-
-                        latest_result = collected_output[-1]
-                        latest_raw = latest_result.get("raw", "")
-
-                        if debug_mode and check_count % 50 == 1:  # Debug every 50 checks (roughly 5 seconds)
-                            self.logger.debug(f"Check #{check_count}, found {len(collected_output)} messages")
-                            self.logger.debug(f"Latest raw (first 100 chars): {repr(latest_raw[:100])}")
-                            self.logger.debug(f"Total content length: {len(all_raw_content)} chars")
-                            print(f"[DEBUG] Check #{check_count}, found {len(collected_output)} messages")
-                            print(f"[DEBUG] Latest raw (first 100 chars): {repr(latest_raw[:100])}")
-                            print(f"[DEBUG] Total content length: {len(all_raw_content)} chars")
-                            # For service ping debugging, show more content
-                            if len(all_raw_content) > 0:
-                                self.logger.debug(f"Service ping content sample: {repr(all_raw_content[:300])}")
-                                print(f"[DEBUG] Service ping content sample: {repr(all_raw_content[:300])}")
-                                if "bytes from" in all_raw_content.lower():
-                                    self.logger.debug("Service ping: Found 'bytes from' pattern")
-                                    print("[DEBUG] Service ping: Found 'bytes from' pattern")
-                                if "seq=" in all_raw_content.lower():
-                                    self.logger.debug("Service ping: Found 'seq=' pattern")
-                                    print("[DEBUG] Service ping: Found 'seq=' pattern")
-                                if "time=" in all_raw_content.lower():
-                                    self.logger.debug("Service ping: Found 'time=' pattern")
-                                    print("[DEBUG] Service ping: Found 'time=' pattern")
-
-                        # Check if ANY of the collected content looks like a final command summary
-                        # Ping completion indicators
-                        ping_indicators = ["round-trip min/avg/max", "round-trip min/avg/max/stddev", "rtt min/avg/max"]
-                        # Service ping completion indicators - SSR service ping specific patterns
-                        service_ping_indicators = [
-                            "service ping completed",  # Generic service ping completion
-                            "service-ping",  # Service ping command reference
-                            "packet transmitted",  # "10 packets transmitted"
-                            "packets transmitted",  # Alternative format
-                            "received",  # "10 received" - common in ping summaries
-                            "packet loss",  # Alternative packet loss indicator
-                            "transmission failure",  # Service ping specific failure
-                            "service path",  # Service path reference
-                            "tenant context",  # Tenant context completion
-                            "service route",  # Service routing completion
-                        ]
-                        # ARP completion indicators - specific patterns from actual ARP output
-                        arp_indicators = [
-                            "total mac entries",  # "Total 31 MAC Entries."
-                            "total flows:",  # "Total Flows:151"
-                            "mac-flow hi-water",  # "Mac-Flow Hi-Water:2865"
-                            "arp table",  # Generic ARP table reference
-                            "no arp entries",  # Empty ARP table
-                            "arp cache",  # ARP cache reference
-                        ]
-                        # Gateway-specific completion indicators (SSR gateways often have shorter output)
-                        gateway_indicators = [
-                            "connected routes",  # Gateway routing table
-                            "total entries",  # Gateway route totals
-                            "kernel routes",  # SSR kernel routing
-                            "bgp routes",  # BGP routing information
-                            "static routes",  # Static route information
-                            "route table",  # Generic route table reference
-                        ]
-                        # Switch-specific completion indicators
-                        switch_indicators = [
-                            "learning table",  # Switch MAC learning
-                            "fdb entries",  # Forwarding database
-                            "vlan information",  # VLAN details
-                            "port statistics",  # Port stats
-                            "interface status",  # Interface information
-                            "ethernet switching table",  # MAC table header
-                            "entries, 40 learned",  # MAC table summary (specific count varies)
-                            "entries,",  # Generic MAC table entry count
-                            "learned",  # MAC learning completion
-                        ]
-                        # General completion indicators
-                        general_indicators = ["command completed", "operation complete", "finished"]
-
-                        all_indicators = (
-                            ping_indicators
-                            + service_ping_indicators
-                            + arp_indicators
-                            + gateway_indicators
-                            + switch_indicators
-                            + general_indicators
-                        )
-                        found_indicator = None
-
-                        if (
-                            debug_mode and check_count % 100 == 1
-                        ):  # Debug indicator checking every 100 checks (roughly 10 seconds)
-                            self.logger.debug(f"Checking {len(all_indicators)} completion indicators")
-                            self.logger.debug(
-                                f"Content sample for indicator check: {repr(all_raw_content.lower()[:150])}"
-                            )
-                            print(f"[DEBUG] Checking {len(all_indicators)} completion indicators")
-                            print(f"[DEBUG] Content sample for indicator check: {repr(all_raw_content.lower()[:150])}")
-
-                        for indicator in all_indicators:
-                            # Skip generic indicators for MAC table commands to allow proper completion detection
-                            if "ethernet switching table" in all_raw_content.lower() and indicator in [
-                                "ethernet switching table",
-                                "entries,",
-                                "learned",
-                            ]:
-                                continue  # Let MAC table specific logic handle this
-                            if indicator in all_raw_content.lower():
-                                found_indicator = indicator
-                                if debug_mode:
-                                    self.logger.debug(f"FOUND completion indicator: '{indicator}'")
-                                    print(f"[DEBUG] FOUND completion indicator: '{indicator}'")
-                                break
-
-                        # Alternative completion: look for "packet loss" followed by "round-trip" pattern (ping specific)  # noqa: E501
-                        if not found_indicator and "packet loss" in all_raw_content.lower():
-                            # Check if we have the complete statistics block
-                            lines = all_raw_content.lower().split("\n")
-                            for line in lines:
-                                if "packet loss" in line and (
-                                    "round-trip" in all_raw_content.lower() or "rtt" in all_raw_content.lower()
-                                ):
-                                    found_indicator = "complete statistics block"
-                                    if debug_mode:
-                                        self.logger.debug("FOUND ping statistics completion pattern")
-                                        self.logger.debug(f"Packet loss line: {repr(line[:100])}")
-                                        print("[DEBUG] FOUND ping statistics completion pattern")
-                                        print(f"[DEBUG] Packet loss line: {repr(line[:100])}")
-                                    break
-
-                        # Service ping specific completion: look for individual ping responses with timing
-                        if (
-                            not found_indicator and len(collected_output) >= 3
-                        ):  # Service ping typically has multiple responses
-                            # Look for service ping patterns - individual responses with seq/ttl/time
-                            service_ping_pattern_count = 0
-                            if "seq=" in all_raw_content.lower() and (
-                                "ttl=" in all_raw_content.lower() or "time=" in all_raw_content.lower()
-                            ):
-                                service_ping_pattern_count += 1
-                            if "bytes from" in all_raw_content.lower():
-                                service_ping_pattern_count += 1
-
-                            if debug_mode and check_count % 200 == 1:  # Debug service ping patterns
-                                self.logger.debug(
-                                    f"Service ping pattern analysis: found {service_ping_pattern_count} service ping indicators"  # noqa: E501
-                                )
-                                print(
-                                    f"[DEBUG] Service ping pattern analysis: found {service_ping_pattern_count} service ping indicators"  # noqa: E501
-                                )
-                                if "seq=" in all_raw_content.lower():
-                                    self.logger.debug("Found seq= pattern in service ping output")
-                                    print("[DEBUG] Found seq= pattern in service ping output")
-                                if "bytes from" in all_raw_content.lower():
-                                    self.logger.debug("Found 'bytes from' pattern in service ping output")
-                                    print("[DEBUG] Found 'bytes from' pattern in service ping output")
-
-                            # If we see service ping patterns and have been collecting for reasonable time
-                            if service_ping_pattern_count >= 2:
-                                # For service ping, if we have multiple ping responses and some idle time, consider complete  # noqa: E501
-                                if (
-                                    time.time() - last_activity > 3
-                                ):  # Wait 3 seconds after last response for service ping
-                                    found_indicator = "service ping pattern detected"
-                                    if debug_mode:
-                                        self.logger.debug(
-                                            f"FOUND service ping completion: {service_ping_pattern_count} patterns detected"  # noqa: E501
-                                        )
-                                        self.logger.debug(f"Service ping idle time: {time.time() - last_activity:.1f}s")
-                                        print(
-                                            f"[DEBUG] FOUND service ping completion: {service_ping_pattern_count} patterns detected"  # noqa: E501
-                                        )
-                                        print(f"[DEBUG] Service ping idle time: {time.time() - last_activity:.1f}s")
-
-                        # Alternative service ping completion: check for count-based completion
-                        if not found_indicator and len(collected_output) >= 5:  # Reasonable number of responses
-                            # Count individual ping responses in format: "64 bytes from X.X.X.X: seq=N ttl=N time=N ms"
-                            ping_response_count = all_raw_content.lower().count("bytes from")
-                            if (
-                                ping_response_count >= 5 and time.time() - last_activity > 2
-                            ):  # Have responses and idle time
-                                found_indicator = f"count-based completion ({ping_response_count} responses)"
-                                if debug_mode:
-                                    self.logger.debug(
-                                        f"FOUND count-based service ping completion: {ping_response_count} responses"
-                                    )
-                                    self.logger.debug(
-                                        f"Idle time since last response: {time.time() - last_activity:.1f}s"
-                                    )
-                                    print(
-                                        f"[DEBUG] FOUND count-based service ping completion: {ping_response_count} responses"  # noqa: E501
-                                    )
-                                    print(f"[DEBUG] Idle time since last response: {time.time() - last_activity:.1f}s")
-
-                        # MAC table completion: detect when table is complete and device stops sending
-                        if not found_indicator and (
-                            "ethernet switching table" in all_raw_content.lower()
-                            or "thernet switching table" in all_raw_content.lower()
-                        ):
-                            # Search for "Ethernet switching table : XXX entries" pattern in reassembled buffer
-                            import re
-
-                            # Look for pattern like "Ethernet switching table : 44 entries" (handles chunking)
-                            table_pattern = r"ethernet switching table\s*:\s*(\d+)\s+entries"
-                            match = re.search(table_pattern, all_raw_content.lower())
-
-                            if match:
-                                entry_count = int(match.group(1))
-
-                                # First check: if we're getting the same message content repeatedly (completion signal)
-                                if len(collected_output) >= 5:
-                                    # Get the last 5 message contents
-                                    last_messages = [msg.get("raw", "") for msg in collected_output[-5:]]
-                                    # If all 5 are identical (and not empty), the command has finished
-                                    if len(set(last_messages)) == 1 and last_messages[0].strip():
-                                        found_indicator = f"mac table completion (detected {len(last_messages)} repeated identical messages)"  # noqa: E501
-                                        if debug_mode:
-                                            self.logger.debug(
-                                                f"FOUND MAC table completion: {len(last_messages)} repeated identical messages detected"  # noqa: E501
-                                            )
-                                            self.logger.debug(f"Repeated message: {repr(last_messages[0][:100])}")
-                                            print(
-                                                f"[DEBUG] FOUND MAC table completion: {len(last_messages)} repeated identical messages detected"  # noqa: E501
-                                            )
-                                            print(f"[DEBUG] Repeated message: {repr(last_messages[0][:100])}")
-                                    else:
-                                        if debug_mode and check_count % 50 == 1:
-                                            # Show how many unique messages in the last 5
-                                            unique_count = len(set(last_messages))
-                                            print(
-                                                f"[DEBUG] MAC table: found {entry_count} entries, last 5 messages have {unique_count} unique contents"  # noqa: E501
-                                            )
-
-                                # Second check: if device has been idle for 3+ seconds and we have substantial MAC entries  # noqa: E501
-                                if not found_indicator and len(collected_output) >= 10 and entry_count >= 10:
-                                    idle_time = time.time() - last_activity
-                                    if idle_time >= 3.0:  # Device has been quiet for 3 seconds
-                                        found_indicator = f"mac table completion (idle timeout: {entry_count} entries, {idle_time:.1f}s idle)"  # noqa: E501
-                                        if debug_mode:
-                                            self.logger.debug(
-                                                f"FOUND MAC table completion via idle timeout: {entry_count} entries, {idle_time:.1f}s idle"  # noqa: E501
-                                            )
-                                            print(
-                                                f"[DEBUG] FOUND MAC table completion via idle timeout: {entry_count} entries, {idle_time:.1f}s idle"  # noqa: E501
-                                            )
-
-                                if not found_indicator and debug_mode and check_count % 50 == 1:
-                                    idle_time = time.time() - last_activity
-                                    print(f"[DEBUG] MAC table: found {entry_count} entries, idle for {idle_time:.1f}s")
-                            else:
-                                if debug_mode and check_count % 50 == 1:
-                                    print(
-                                        f"[DEBUG] MAC table: checking for completion pattern in {len(all_raw_content)} chars"  # noqa: E501
-                                    )
-
-                        # ARP-specific completion: check for structured ARP output patterns
-                        if not found_indicator and len(collected_output) >= 2:
-                            # Look for ARP table structure patterns
-                            arp_patterns = ["ip address", "hw address", "interface", "incomplete", "permanent"]
-                            arp_pattern_count = sum(1 for pattern in arp_patterns if pattern in all_raw_content.lower())
-
-                            if debug_mode and check_count % 200 == 1:  # Debug ARP patterns less frequently
-                                print(
-                                    f"[DEBUG] ARP pattern analysis: found {arp_pattern_count}/{len(arp_patterns)} patterns"  # noqa: E501
-                                )
-                                found_patterns = [p for p in arp_patterns if p in all_raw_content.lower()]
-                                print(f"[DEBUG] Found ARP patterns: {found_patterns}")
-
-                            # If we see multiple ARP patterns, this might be a complete ARP table
-                            if arp_pattern_count >= 2:
-                                # Check if we've been collecting for at least 1 second (ARP commands are usually fast)
-                                if time.time() - last_activity > 1:
-                                    found_indicator = "arp table structure detected"
-                                    if debug_mode:
-                                        print(
-                                            f"[DEBUG] FOUND ARP table completion: {arp_pattern_count} patterns detected"
-                                        )
-
-                        if found_indicator:
-                            # This appears to be the final ping summary
-                            if debug_mode:
-                                self.logger.debug(f"Found completion indicator '{found_indicator}' in combined content")
-                                self.logger.debug(f"Completing after {check_count} checks")
-                                self.logger.debug(f"Total collected messages: {len(collected_output)}")
-                                self.logger.debug(f"Total content length: {len(all_raw_content)} characters")
-                                self.logger.debug(
-                                    f"Raw content sample (first 200 chars): {repr(all_raw_content[:200])}"
-                                )
-                                self.logger.debug(
-                                    f"Raw content sample (last 200 chars): {repr(all_raw_content[-200:])}"
-                                )
-                                print(f"[DEBUG] Found completion indicator '{found_indicator}' in combined content")
-                                print(f"[DEBUG] Completing after {check_count} checks")
-                                print(f"[DEBUG] Total collected messages: {len(collected_output)}")
-                                print(f"[DEBUG] Total content length: {len(all_raw_content)} characters")
-                                print(f"[DEBUG] Raw content sample (first 200 chars): {repr(all_raw_content[:200])}")
-                                print(f"[DEBUG] Raw content sample (last 200 chars): {repr(all_raw_content[-200:])}")
-                            final_results = self.command_results.pop(session_id)
-                            break
-                else:
-                    if debug_mode and check_count % 50 == 1:  # Debug every 5 seconds
-                        self.logger.debug(f"Check #{check_count}, no results yet for session {session_id}")
-                        self.logger.debug(f"Available sessions: {list(self.command_results.keys())}")
-                        print(f"[DEBUG] Check #{check_count}, no results yet for session {session_id}")
-                        print(f"[DEBUG] Available sessions: {list(self.command_results.keys())}")
-
-            # Emergency circuit breaker - if we're doing too many checks, something is wrong
-            if check_count > 10000:  # At 0.1s per check, this is ~16 minutes
-                if debug_mode:
-                    self.logger.error(f"Circuit breaker triggered at {check_count} checks!")
-                    self.logger.error("This indicates a possible infinite loop or system hang")
-                    print(f"[EMERGENCY] Circuit breaker triggered at {check_count} checks!")
-                    print("[EMERGENCY] This indicates a possible infinite loop or system hang")
-                self.logger.error(f"Emergency circuit breaker: {check_count} checks exceeded for session {session_id}")
-                with self.results_lock:
-                    final_results = self.command_results.pop(session_id, [])
-                return final_results if final_results else None
-
-            # Check for activity timeout (no new messages)
-            collected_count = 0
-            with self.results_lock:
-                if session_id in self.command_results:
-                    collected_count = len(self.command_results[session_id])
-
-            if collected_count > 0 and (time.time() - last_activity > activity_timeout):
-                if debug_mode:
-                    self.logger.debug(
-                        f"Activity timeout reached ({activity_timeout}s), completing with {collected_count} messages"
-                    )
-                    print(
-                        f"[DEBUG] Activity timeout reached ({activity_timeout}s), completing with {collected_count} messages"  # noqa: E501
-                    )
-                self.logger.info(f"No new data for {activity_timeout}s, assuming command complete")
-                with self.results_lock:
-                    if session_id in self.command_results:
-                        final_results = self.command_results.pop(session_id)
-                        break
-
-            # Critical: Ensure we don't create a busy wait loop
-            time.sleep(0.1)  # Check every 100ms - DO NOT REMOVE THIS SLEEP
-        else:
-            # Timeout occurred
-            if debug_mode:
-                self.logger.debug(f"Timeout occurred after {timeout_seconds}s, {check_count} checks")
-                print(f"[DEBUG] Timeout occurred after {timeout_seconds}s, {check_count} checks")
-            with self.results_lock:
-                final_results = self.command_results.pop(session_id, [])
-
-            if not final_results:
-                if debug_mode:
-                    print(f"[DEBUG] No results collected for session {session_id}")
-                self.logger.warning(f"Timeout waiting for command result: {session_id}")
-                perf_monitor.finish()  # type: ignore[no-untyped-call]  # Mark performance monitoring as complete
-                return None
-
-        # Combine all collected output
-        perf_monitor.finish()  # type: ignore[no-untyped-call]  # Mark performance monitoring as complete
-
-        if final_results:
-            if debug_mode:
-                self.logger.debug(f"Combining {len(final_results)} result segments")
-                self.logger.debug(f"Total wait time: {time.time() - start_time:.2f} seconds")
-                self.logger.debug(f"Total checks performed: {check_count}")
-                print(f"[DEBUG] Combining {len(final_results)} result segments")
-                print(f"[DEBUG] Total wait time: {time.time() - start_time:.2f} seconds")
-                print(f"[DEBUG] Total checks performed: {check_count}")
-
-            combined_raw = ""
-            combined_other: dict[str, Any] = {}
-
-            for index, result in enumerate(final_results):
-                raw_content = result.get("raw", "")
-                if raw_content:
-                    combined_raw += raw_content
-                    if debug_mode and len(final_results) > 5:  # Only show details for complex results
-                        print(f"[DEBUG] Segment {index + 1}: {len(raw_content)} chars")
-
-                # Collect any other fields
-                for key, value in result.items():
-                    if key not in ["raw", "session"]:
-                        if key in combined_other:
-                            combined_other[key] = str(combined_other[key]) + str(value)
-                        else:
-                            combined_other[key] = value
-
-            # Return combined result
-            final_result = {"raw": combined_raw, "session": session_id}
-            final_result.update(combined_other)
-
-            if debug_mode:
-                print(f"[DEBUG] Final combined result length: {len(combined_raw)} characters")
-                print(f"[DEBUG] Final result fields: {list(final_result.keys())}")
-                print(f"[DEBUG] First 150 chars of final result: {repr(combined_raw[:150])}")
-                print(f"[DEBUG] Last 150 chars of final result: {repr(combined_raw[-150:])}")
-                if len(combined_raw) == 0:
-                    print("[DEBUG] WARNING: Final result is empty - this may indicate an issue")
-                print(f"[DEBUG] Session {session_id} result collection complete")
-                print("[DEBUG] " + "=" * 60)
-
-            self.logger.info(f"Command completed with {len(final_results)} message segments")
-            return final_result
-
-        perf_monitor.finish()  # type: ignore[no-untyped-call]  # Mark performance monitoring as complete
-        return None
-
-    def _on_open(self, websocket_connection):  # type: ignore[no-untyped-def]
-        """WebSocket connection opened callback."""
-        self.connected = True
-        self.logger.debug("WebSocket connection opened")
-
-    def _on_message(self, websocket_connection, message):  # type: ignore[no-untyped-def]  # noqa: C901, PLR0912, PLR0915
-        """
-        WebSocket message received callback.
-
-        Processes incoming messages following the documented Mist API format:
-        {
-            "event": "data",
-            "channel": "/sites/{site_id}/devices/{device_id}/cmd",
-            "data": {
-                "session": "session_id",
-                "raw": "64 bytes from 23.211.0.110: seq=8 ttl=58 time=12.323 ms\n"
-            }
-        }
-        """
-        debug_mode = is_debug_mode()  # type: ignore[no-untyped-call]
-
-        try:
-            if debug_mode:
-                print(f"[DEBUG] Raw WebSocket message received: {repr(message)} (type: {type(message)})")
-            self.logger.debug(f"Raw WebSocket message received: {repr(message)} (type: {type(message)})")
-
-            # Parse JSON message - handle string messages first
-            message_data = None
-            if isinstance(message, str):
-                try:
-                    message_data = json.loads(message)
-                    if debug_mode:
-                        print(f"[DEBUG] Successfully parsed JSON message: {message_data}")
-                    self.logger.debug(f"Successfully parsed JSON message: {message_data}")
-                except json.JSONDecodeError as json_error:
-                    if debug_mode:
-                        print(f"[DEBUG] Failed to parse JSON message: {json_error}")
-                        print(f"[DEBUG] Raw message content: {repr(message)}")
-                    self.logger.warning(f"Failed to parse JSON message: {json_error}")
-                    self.logger.debug(f"Raw message content: {repr(message)}")
-                    return
-            elif isinstance(message, dict):
-                message_data = message
-                if debug_mode:
-                    print(f"[DEBUG] Received dict message: {message_data}")
-                self.logger.debug(f"Received dict message: {message_data}")
-            else:
-                if debug_mode:
-                    print(f"[DEBUG] Unexpected message type: {type(message)}, content: {repr(message)}")
-                self.logger.warning(f"Unexpected message type: {type(message)}, content: {repr(message)}")
-                return
-
-            # Ensure we have a valid message_data dict before proceeding
-            if not isinstance(message_data, dict):
-                if debug_mode:
-                    print(f"[DEBUG] Message data is not a dict after parsing: {type(message_data)}")
-                self.logger.error(f"Message data is not a dict after parsing: {type(message_data)}")
-                return
-
-            # Enhanced packet content logging for debug mode
-            if debug_mode:
-                print("[PACKET] WebSocket packet details:")
-                print(f"[PACKET]   Event: {message_data.get('event', 'unknown')}")
-                print(f"[PACKET]   Channel: {message_data.get('channel', 'unknown')}")
-                if "data" in message_data:
-                    data_content = message_data["data"]
-                    print(f"[PACKET]   Data type: {type(data_content)}")
-                    if isinstance(data_content, dict):
-                        print(f"[PACKET]   Data keys: {list(data_content.keys())}")
-                        if "session" in data_content:
-                            session_id = data_content["session"]
-                            print(f"[PACKET]   Session ID: {session_id}")
-                        if "raw" in data_content:
-                            raw_content = data_content["raw"]
-                            print(f"[PACKET]   Raw content length: {len(str(raw_content))} chars")
-                            print(f"[PACKET]   Raw content: {repr(raw_content)}")
-                    else:
-                        print(f"[PACKET]   Data content: {repr(data_content)}")
-                else:
-                    print("[PACKET]   No data field in message")
-
-            # Handle subscription confirmation
-            if message_data.get("event") == "channel_subscribed":
-                channel = message_data.get("channel")
-                if debug_mode:
-                    print(f"[DEBUG] Channel subscription confirmed: {channel}")
-                self.logger.info(f"Channel subscription confirmed: {channel}")
-                # Track confirmed subscription
-                if channel:
-                    self.confirmed_subscriptions.add(channel)
-                return
-
-            # Handle command data following documented format
-            if message_data.get("event") == "data":
-                channel = message_data.get("channel", "")
-                data_payload = message_data.get("data", {})
-
-                if debug_mode:
-                    print(f"[DEBUG] Processing data event from channel: {channel}")
-                    print(f"[DEBUG] Data payload type: {type(data_payload)}")
-                    print(f"[DEBUG] Data payload content: {repr(data_payload)}")
-
-                # Check if data_payload is actually a nested message structure
-                if isinstance(data_payload, str):
-                    try:
-                        # Sometimes the data field contains a JSON string
-                        data_payload = json.loads(data_payload)
-                        if debug_mode:
-                            print(f"[DEBUG] Parsed nested JSON in data field: {data_payload}")
-                        self.logger.debug(f"Parsed nested JSON in data field: {data_payload}")
-                    except json.JSONDecodeError:
-                        if debug_mode:
-                            print(f"[DEBUG] Data field is string but not JSON: {data_payload}")
-                        self.logger.warning(f"Data field is string but not JSON: {data_payload}")
-                        return
-
-                # Check if we have another nested event structure
-                if isinstance(data_payload, dict) and data_payload.get("event") == "data":
-                    # This is a nested structure, extract the actual data
-                    actual_data = data_payload.get("data", {})
-                    if debug_mode:
-                        print(f"[DEBUG] Found nested event structure, extracting actual data: {actual_data}")
-                    self.logger.debug(f"Found nested event structure, extracting actual data: {actual_data}")
-                    data_payload = actual_data
-
-                session_id = data_payload.get("session") if isinstance(data_payload, dict) else None
-
-                if debug_mode:
-                    print(f"[DEBUG] Processing data event - channel: {channel}, session: {session_id}")
-                    print(f"[DEBUG] Final data payload: {data_payload}")
-                    if session_id:
-                        print(f"[DEBUG] Session ID extracted: {session_id}")
-                    else:
-                        print("[DEBUG] No session ID found in data payload")
-
-                self.logger.debug(f"Processing data event - channel: {channel}, session: {session_id}")
-                self.logger.debug(f"Final data payload: {data_payload}")
-
-                if session_id:
-                    # Store each message for streaming commands like ping
-                    with self.results_lock:
-                        # Initialize a list for this session if it doesn't exist
-                        if session_id not in self.command_results:
-                            self.command_results[session_id] = []
-                            if debug_mode:
-                                print(f"[DEBUG] Initialized new result list for session: {session_id}")
-                        # Append this message to the list
-                        self.command_results[session_id].append(data_payload)
-
-                        if debug_mode:
-                            current_count = len(self.command_results[session_id])
-                            print(f"[DEBUG] Stored message #{current_count} for session {session_id}")
-                            if "raw" in data_payload:
-                                raw_data = data_payload["raw"]
-                                print(f"[DEBUG] Raw data in stored message: {repr(raw_data)}")
-                            print(f"[DEBUG] Complete stored message: {data_payload}")
-
-                        self.logger.debug(f"Stored command result for session {session_id}: {data_payload}")
-                    self.logger.debug(f"Command result segment received for session: {session_id}")
-                    self.logger.debug(f"Total segments for session: {len(self.command_results[session_id])}")
-                else:
-                    self.logger.warning(f"Received data event without session ID. Full message: {message_data}")
-                    self.logger.warning(f"Data payload: {data_payload}")
-            else:
-                self.logger.debug(f"Unhandled message event type: {message_data.get('event')}")
-
-        except Exception as message_error:
-            self.logger.error(f"Error processing WebSocket message: {message_error}")
-            self.logger.debug("Exception details:", exc_info=True)
-            self.logger.debug(f"Problematic message: {repr(message)}")
-            self.logger.debug(f"Message type: {type(message)}")
-
-    def _on_error(self, websocket_connection, error):  # type: ignore[no-untyped-def]
-        """WebSocket error callback."""
-        self.logger.debug(f"WebSocket error type: {type(error).__name__}")
-        self.logger.error(f"WebSocket error: {error}")
-
-    def _on_close(self, websocket_connection, close_status_code, close_message):  # type: ignore[no-untyped-def]
-        """WebSocket connection closed callback."""
-        self.connected = False
-        self.logger.debug(f"WebSocket close details: status_code={close_status_code}, message={close_message}")
-        self.logger.info(f"WebSocket connection closed (status: {close_status_code})")
-
-    def disconnect(self) -> None:
-        """Close WebSocket connection and cleanup resources."""
-        if self.websocket_connection:
-            self.websocket_connection.close()
-        self.connected = False
-        self.subscribed_channels.clear()
-
-        with self.results_lock:
-            self.command_results.clear()
 
 
 class PacketCaptureManager:
@@ -18034,948 +17192,6 @@ class SiteExportUtils:
             all_sites_fn=APICoreFetchUtils.all_sites_with_limit,
             save_data_fn=DataExporter.save_data_to_output,
         )
-
-
-class WebSocketNetworkDiagCommands:
-    """
-    WebSocket Network Diagnostic Commands
-
-    Handles ping, ARP, and service ping operations via WebSocket.
-    Extracted from WebSocketCommands.
-    """
-
-    @staticmethod
-    def ping_device():  # type: ignore[no-untyped-def]  # noqa: C901, PLR0912, PLR0915
-        """
-        Execute ping command on a network device via WebSocket.
-
-        Follows the documented Mist API pattern:
-        1. Connect to WebSocket
-        2. Subscribe to device command channel
-        3. Issue POST ping command
-        4. Await results via WebSocket stream
-
-        SECURITY: Uses authenticated WebSocket connection with session-based
-        command demultiplexing for concurrent command safety.
-        """
-        logging.info("Starting WebSocket ping operation...")
-
-        # Check for debug mode from command line args
-        debug_mode = "--debug" in sys.argv or "-d" in sys.argv
-
-        if debug_mode:
-            logging.getLogger().setLevel(logging.DEBUG)
-            print("[DEBUG] DEBUG MODE ENABLED")
-
-        logging.info("Starting WebSocket ping operation...")
-        logging.debug("ENTER: ping_device_websocket")
-
-        try:
-            # Interactive site and device selection
-            site_id = PromptUtils.select_site_id_from_csv()
-            if not site_id:
-                print("! No site selected. Operation cancelled.")
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Selected site_id = {site_id}")
-
-            # Get device selection
-            device_id = PromptUtils.select_device_id_from_inventory(site_id, device_type="all")
-            if not device_id:
-                print("! No device selected. Operation cancelled.")
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Selected device_id = {device_id}")
-
-            # Get ping target from user (default to 8.8.8.8)
-            target_input = input("Enter the target hostname or IP address to ping (default: 8.8.8.8): ").strip()
-            target_host = target_input if target_input else "8.8.8.8"
-
-            # Validate target host
-            if not ValidationUtils.validate_ping_target(target_host):
-                print(f"! Invalid ping target: {target_host}")
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Target host = {target_host}")
-
-            # Get ping count (optional)
-            ping_count_input = input("Enter number of ping packets (default: 4): ").strip()
-            ping_count = 4
-            if ping_count_input:
-                try:
-                    ping_count = int(ping_count_input)
-                    if ping_count < 1 or ping_count > 100:
-                        print("! Ping count must be between 1 and 100. Using default: 4")
-                        ping_count = 4
-                except ValueError:
-                    print("! Invalid ping count. Using default: 4")
-                    ping_count = 4
-
-            if debug_mode:
-                print(f"[DEBUG] Ping count = {ping_count}")
-
-            print(f"\n-> Executing ping to {target_host} on device {device_id}...")
-            print(f"-> Ping count: {ping_count}")
-            print("-> Establishing WebSocket connection...")
-
-            # Initialize WebSocket manager
-            websocket_manager = WebSocketManager(apisession)
-
-            if debug_mode:
-                print("[DEBUG] WebSocketManager initialized")
-
-            # Connect to WebSocket
-            if not websocket_manager.connect():
-                print("! Failed to establish WebSocket connection")
-                return
-
-            if debug_mode:
-                print("[DEBUG] WebSocket connection established")
-
-            # Subscribe to device command channel
-            command_channel = f"/sites/{site_id}/devices/{device_id}/cmd"
-            if not websocket_manager.subscribe_to_channel(command_channel):
-                print("! Failed to subscribe to device command channel")
-                websocket_manager.disconnect()
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Subscribed to channel: {command_channel}")
-
-            print("-> WebSocket connected and subscribed")
-
-            # Wait a moment for subscription to be established
-            time.sleep(1)
-
-            # Issue ping command via REST API
-            ping_payload = {"host": target_host, "count": ping_count}
-
-            print("-> Issuing ping command...")
-            logging.debug(f"Ping payload: {ping_payload}")
-
-            if debug_mode:
-                print(f"[DEBUG] Ping payload = {ping_payload}")
-
-            # Get authentication details for direct HTTP request
-            mist_host = getattr(apisession, "host", None) or os.getenv("MIST_HOST")
-            mist_apitoken = getattr(apisession, "apitoken", None) or os.getenv("MIST_APITOKEN")
-
-            if not mist_host or not mist_apitoken:
-                print("! Mist host or API token not found in session or environment")
-                websocket_manager.disconnect()
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] mist_host = {mist_host}")
-                print(f"[DEBUG] API token length = {len(mist_apitoken) if mist_apitoken else 0}")
-
-            # Make direct POST request to trigger ping
-            ping_url = f"https://{mist_host}/api/v1/sites/{site_id}/devices/{device_id}/ping"
-            headers = {"Authorization": f"Token {mist_apitoken}", "Content-Type": "application/json"}
-
-            if debug_mode:
-                print(f"[DEBUG] POST URL = {ping_url}")
-                print("[DEBUG] Headers = {'Authorization': 'Token [REDACTED]', 'Content-Type': 'application/json'}")
-
-            ping_response = requests.post(ping_url, headers=headers, json=ping_payload, timeout=30)
-
-            if debug_mode:
-                print(f"[DEBUG] HTTP Response Status = {ping_response.status_code}")
-                print(f"[DEBUG] HTTP Response Body = {ping_response.text}")
-
-            if ping_response.status_code != 200:
-                print(f"! Failed to issue ping command: {ping_response.status_code}")
-                print(f"! Response: {ping_response.text}")
-                websocket_manager.disconnect()
-                return
-
-            # Extract session ID from response
-            response_data = ping_response.json()
-            session_id = response_data.get("session")
-            if not session_id:
-                print("! No session ID returned from ping command")
-                websocket_manager.disconnect()
-                return
-
-            print(f"-> Ping command issued (session: {session_id[:8]}...)")
-            print("-> Waiting for ping results...")
-
-            if debug_mode:
-                print(f"[DEBUG] Full session ID = {session_id}")
-                print("[DEBUG] Starting to wait for WebSocket results...")
-
-            # Wait for ping results via WebSocket
-            ping_result = websocket_manager.wait_for_command_result(session_id, timeout_seconds=30)
-
-            if debug_mode:
-                print(f"[DEBUG] wait_for_command_result returned: {ping_result is not None}")
-                if ping_result:
-                    print(f"[DEBUG] Result keys: {list(ping_result.keys())}")
-
-            if ping_result:
-                print("\n" + "=" * 60)
-                print("PING RESULTS:")
-                print("=" * 60)
-
-                # Display raw output (this is where ping results come according to documentation)
-                raw_output = ping_result.get("raw", "")
-                if raw_output:
-                    print("RAW OUTPUT:")
-                    print("-" * 40)
-                    print(raw_output)
-
-                # Display any other output fields that might be present
-                output_fields = ping_result.get("Output", "")
-                if output_fields and output_fields != raw_output:
-                    print("\nOTHER OUTPUT:")
-                    print("-" * 40)
-                    print(output_fields)
-
-                # Show all available fields for debugging
-                available_fields = [key for key in ping_result.keys() if key not in ["raw", "Output", "session"]]
-                if available_fields:
-                    print(f"\nOTHER AVAILABLE FIELDS: {available_fields}")
-                    for field in available_fields:
-                        field_value = ping_result.get(field)
-                        if field_value:
-                            print(f"{field}: {field_value}")
-
-                if not raw_output and not output_fields:
-                    print("No output data received")
-                    print(f"Available result keys: {list(ping_result.keys())}")
-
-                print("=" * 60)
-
-                # Log the successful operation
-                logging.info(f"WebSocket ping completed successfully for {target_host}")
-
-            else:
-                print("! Timeout waiting for ping results")
-                logging.warning("WebSocket ping operation timed out")
-
-                if debug_mode:
-                    print("[DEBUG] Checking WebSocket manager state...")
-                    print(f"[DEBUG] Connected = {websocket_manager.connected}")
-                    print(f"[DEBUG] Subscribed channels = {websocket_manager.subscribed_channels}")
-                    with websocket_manager.results_lock:
-                        print(f"[DEBUG] Pending results = {list(websocket_manager.command_results.keys())}")
-
-        except Exception as ping_error:
-            error_message = f"WebSocket ping operation failed: {ping_error}"
-            print(f"! {error_message}")
-            logging.error(error_message)
-
-            if debug_mode:
-                print("[DEBUG] Exception details:")
-                import traceback
-
-                traceback.print_exc()
-
-            logging.debug("EXIT: ping_device_websocket - error")
-
-        finally:
-            # Always cleanup WebSocket connection
-            try:
-                websocket_manager_local = locals().get("websocket_manager")
-                if websocket_manager_local is not None:
-                    websocket_manager_local.disconnect()
-                    print("-> WebSocket connection closed")
-
-                    if debug_mode:
-                        print("[DEBUG] WebSocket cleanup completed")
-            except Exception as cleanup_error:
-                logging.warning(f"WebSocket cleanup error: {cleanup_error}")
-
-            logging.debug("EXIT: ping_device_websocket")
-
-    @staticmethod
-    def arp_device():  # type: ignore[no-untyped-def]  # noqa: C901, PLR0912, PLR0915
-        """
-        Execute ARP command on a network device via WebSocket.
-
-        Follows the documented Mist API pattern for ARP commands:
-        1. Subscribe to WebSocket channel
-        2. POST ARP command
-        3. Receive results via WebSocket stream with session-based demultiplexing
-
-        SECURITY: Uses authenticated WebSocket connection with session-based
-        command demultiplexing for concurrent command safety.
-        """
-        logging.info("Starting WebSocket ARP operation...")
-        logging.debug("ENTER: arp_device_websocket")
-
-        debug_mode = "--debug" in sys.argv or "-d" in sys.argv
-
-        if debug_mode:
-            print("[DEBUG] Starting ARP via WebSocket operation...")
-
-        try:
-            # Interactive site and device selection
-            site_id = PromptUtils.select_site_id_from_csv()
-            if not site_id:
-                print("! No site selected. Operation cancelled.")
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Selected site_id = {site_id}")
-
-            # Get device selection
-            device_id = PromptUtils.select_device_id_from_inventory(site_id, device_type="all")
-            if not device_id:
-                print("! No device selected. Operation cancelled.")
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Selected device_id = {device_id}")
-
-            # Get device details to check type and model for compatibility
-            device_info = None
-            try:
-                rawdata = mistapi.api.v1.sites.devices.listSiteDevices(apisession, site_id, type="all").data
-                device_info = next((device for device in rawdata if device.get("id") == device_id), None)
-
-                if device_info:
-                    device_type = device_info.get("type", "unknown")
-                    device_model = device_info.get("model", "unknown")
-                    device_name = device_info.get("name", f"Device {device_id[:8]}")
-
-                    if debug_mode:
-                        print(f"[DEBUG] Device type: {device_type}, model: {device_model}, name: {device_name}")
-
-                    # Warn about device compatibility
-                    if device_type == "switch":
-                        print(f"!?  WARNING: Switch detected (Model: {device_model})")
-                        print("   -> Switches may have limited WebSocket ARP support")
-                        print("   -> Consider using SSH-based ARP commands instead")
-                        print("   -> This operation may timeout or return limited results")
-
-                        response = input("   -> Continue anyway? (y/N): ").strip().lower()
-                        if response not in ["y", "yes"]:
-                            print("! Operation cancelled by user")
-                            return
-
-                    elif device_type == "gateway":
-                        print(f"!? Gateway detected (Model: {device_model})")
-                        print("   -> Gateways have good WebSocket ARP support")
-                        print("   -> Results may differ from Access Points")
-
-                    elif device_type == "ap":
-                        print(f"!? Access Point detected (Model: {device_model})")
-                        print("   -> Access Points have full WebSocket ARP support")
-
-                    else:
-                        print(f"? Unknown device type: {device_type} (Model: {device_model})")
-                        print("   -> Proceeding with standard ARP command")
-
-            except Exception as device_check_error:
-                logging.warning(f"Could not verify device compatibility: {device_check_error}")
-                if debug_mode:
-                    print(f"[DEBUG] Device check failed: {device_check_error}")
-                print("   -> Proceeding with standard ARP command")
-
-            print(f"\n-> Executing ARP command on device {device_id}...")
-            print("-> Establishing WebSocket connection...")
-
-            if debug_mode:
-                print("[DEBUG] WebSocketManager initialized")
-
-            # Initialize WebSocket manager
-            websocket_manager = WebSocketManager(apisession)
-
-            # Connect to WebSocket
-            if not websocket_manager.connect():
-                print("! Failed to establish WebSocket connection")
-                return
-
-            if debug_mode:
-                print("[DEBUG] WebSocket connection established")
-
-            # Subscribe to device command channel
-            command_channel = f"/sites/{site_id}/devices/{device_id}/cmd"
-            if not websocket_manager.subscribe_to_channel(command_channel):
-                print("! Failed to subscribe to device command channel")
-                websocket_manager.disconnect()
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Subscribed to channel: {command_channel}")
-
-            print("-> WebSocket connected and subscribed")
-
-            # Wait a moment for subscription to be established
-            time.sleep(1)
-
-            print("-> Issuing ARP command...")
-
-            # Get authentication details for direct HTTP request
-            mist_host = getattr(apisession, "host", None) or os.getenv("MIST_HOST")
-            mist_apitoken = getattr(apisession, "apitoken", None) or os.getenv("MIST_APITOKEN")
-
-            if debug_mode:
-                print(f"[DEBUG] mist_host = {mist_host}")
-                print(f"[DEBUG] API token length = {len(mist_apitoken) if mist_apitoken else 'None'}")
-
-            if not mist_host or not mist_apitoken:
-                print("! Mist host or API token not found in session or environment")
-                websocket_manager.disconnect()
-                return
-
-            # Make direct POST request to trigger ARP command
-            arp_url = f"https://{mist_host}/api/v1/sites/{site_id}/devices/{device_id}/arp"
-            headers = {"Authorization": f"Token {mist_apitoken}", "Content-Type": "application/json"}
-
-            if debug_mode:
-                print(f"[DEBUG] POST URL = {arp_url}")
-                print("[DEBUG] Headers = {'Authorization': 'Token [REDACTED]', 'Content-Type': 'application/json'}")
-
-            # ARP command typically doesn't need a payload body
-            arp_response = requests.post(arp_url, headers=headers, json={}, timeout=30)
-
-            if debug_mode:
-                print(f"[DEBUG] HTTP Response Status = {arp_response.status_code}")
-                print(f"[DEBUG] HTTP Response Body = {arp_response.text}")
-
-            if arp_response.status_code != 200:
-                print(f"! Failed to issue ARP command: {arp_response.status_code}")
-                print(f"! Response: {arp_response.text}")
-                websocket_manager.disconnect()
-                return
-
-            # Extract session ID from response
-            response_data = arp_response.json()
-            session_id = response_data.get("session")
-            if not session_id:
-                print("! No session ID returned from ARP command")
-                websocket_manager.disconnect()
-                return
-
-            print(f"-> ARP command issued (session: {session_id[:8]}...)")
-            print("-> Waiting for ARP results...")
-
-            if debug_mode:
-                print(f"[DEBUG] Full session ID = {session_id}")
-                print("[DEBUG] Starting to wait for WebSocket results...")
-
-            # Determine timeout based on device type
-            if device_info:
-                device_type = device_info.get("type", "unknown")
-                if device_type == "switch":
-                    # Switches often timeout, give them more time
-                    timeout_seconds = 45
-                    print("   -> Using extended timeout for switch (45 seconds)")
-                elif device_type == "gateway":
-                    # Gateways work but may be slower
-                    timeout_seconds = 35
-                    print("   -> Using extended timeout for gateway (35 seconds)")
-                else:
-                    # APs and unknown devices use standard timeout
-                    timeout_seconds = 30
-            else:
-                timeout_seconds = 30
-
-            # Wait for ARP results via WebSocket
-            arp_result = websocket_manager.wait_for_command_result(session_id, timeout_seconds=timeout_seconds)
-
-            if debug_mode:
-                print(f"[DEBUG] wait_for_command_result returned: {arp_result is not None}")
-                if arp_result:
-                    print(f"[DEBUG] Result keys: {list(arp_result.keys())}")
-
-            if arp_result:
-                print("\n" + "=" * 60)
-                print("ARP TABLE RESULTS:")
-                print("=" * 60)
-
-                # Add device-specific context
-                if device_info:
-                    device_type = device_info.get("type", "unknown")
-                    device_model = device_info.get("model", "unknown")
-                    device_name = device_info.get("name", "Unknown Device")
-
-                    print(f"Device: {device_name} ({device_type.upper()}: {device_model})")
-
-                    if device_type == "switch":
-                        print("Note: Switch ARP data may show forwarding table or limited ARP information")
-                    elif device_type == "gateway":
-                        print("Note: Gateway ARP data may include routing information")
-                    elif device_type == "ap":
-                        print("Note: Access Point ARP data shows client connectivity information")
-
-                    print("-" * 60)
-
-                # Display raw output if available
-                raw_output = arp_result.get("raw", "")
-                if raw_output:
-                    # Parse and display gateway data in table format if it's JSON
-                    if device_info and device_info.get("type") == "gateway" and raw_output.strip().startswith("{"):
-                        try:
-                            # json is already imported globally
-                            gateway_data = json.loads(raw_output)
-
-                            if gateway_data.get("status") == "SUCCESS" and "rows" in gateway_data:
-                                print("PARSED ARP TABLE:")
-                                print("-" * 40)
-
-                                # Get column headers
-                                columns = gateway_data.get("columns", [])
-                                if columns:
-                                    # Create header row
-                                    column_headers = [
-                                        col.get("display_name", col.get("id", "Unknown")) for col in columns
-                                    ]
-
-                                    # Calculate column widths
-                                    rows = gateway_data.get("rows", [])
-                                    col_widths = []
-                                    for idx, header in enumerate(column_headers):
-                                        max_width = len(header)
-                                        for row in rows:
-                                            col_id = columns[idx].get("id", "")
-                                            cell_value = str(row.get(col_id, ""))
-                                            max_width = max(max_width, len(cell_value))
-                                        col_widths.append(min(max_width + 2, 20))  # Cap at 20 chars
-
-                                    # Print header
-                                    header_line = " | ".join(
-                                        header.ljust(col_widths[idx]) for idx, header in enumerate(column_headers)
-                                    )
-                                    print(header_line)
-                                    print("-" * len(header_line))
-
-                                    # Print data rows
-                                    for row in rows:
-                                        row_values = []
-                                        for idx, col in enumerate(columns):
-                                            col_id = col.get("id", "")
-                                            cell_value = str(row.get(col_id, ""))
-                                            # Truncate if too long
-                                            if len(cell_value) > col_widths[idx] - 2:
-                                                cell_value = cell_value[: col_widths[idx] - 5] + "..."
-                                            row_values.append(cell_value.ljust(col_widths[idx]))
-                                        print(" | ".join(row_values))
-
-                                    print(f"\nTotal ARP Entries: {len(rows)}")
-
-                                    # Also show raw for reference if debug mode
-                                    if debug_mode:
-                                        print("\nRAW JSON OUTPUT (Debug):")
-                                        print("-" * 40)
-                                        print(raw_output)
-                                else:
-                                    print("No column information available in gateway response")
-                                    print("RAW OUTPUT:")
-                                    print("-" * 40)
-                                    print(raw_output)
-                            else:
-                                print("Gateway response format not recognized")
-                                print("RAW OUTPUT:")
-                                print("-" * 40)
-                                print(raw_output)
-
-                        except json.JSONDecodeError as json_error:
-                            if debug_mode:
-                                print(f"[DEBUG] Failed to parse gateway JSON: {json_error}")
-                            print("Failed to parse gateway JSON output")
-                            print("RAW OUTPUT:")
-                            print("-" * 40)
-                            print(raw_output)
-                    else:
-                        # For APs and other devices, show raw output
-                        print("RAW OUTPUT:")
-                        print("-" * 40)
-                        print(raw_output)
-
-                # Display parsed output if available
-                parsed_output = arp_result.get("Output", "")
-                if parsed_output and parsed_output != raw_output:
-                    print("\nPARSED OUTPUT:")
-                    print("-" * 40)
-                    print(parsed_output)
-
-                if not raw_output and not parsed_output:
-                    print("No output data received")
-                    if device_info and device_info.get("type") == "switch":
-                        print("\nTroubleshooting for switches:")
-                        print("-> Try using SSH-based commands instead")
-                        print("-> Some switches require specific ARP command syntax")
-                        print("-> WebSocket API may have limited switch support")
-
-                print("=" * 60)
-
-                # Log the successful operation with device context
-                device_context = f"device {device_id}"
-                if device_info:
-                    device_context = f"{device_info.get('type', 'unknown')} {device_info.get('name', device_id[:8])}"
-                logging.info(f"WebSocket ARP completed successfully for {device_context}")
-
-            else:
-                print("! Timeout waiting for ARP results")
-
-                # Provide device-specific troubleshooting advice
-                if device_info:
-                    device_type = device_info.get("type", "unknown")
-                    device_model = device_info.get("model", "unknown")
-
-                    if device_type == "switch":
-                        print(f"\nSwitch troubleshooting ({device_model}):")
-                        print("-> Switches often have limited WebSocket ARP support")
-                        print("-> Try using SSH-based 'show arp' commands instead")
-                        print("-> Some switch models require specific command syntax")
-                        print("-> Consider using Menu option for SSH device commands")
-                    elif device_type == "gateway":
-                        print(f"\nGateway troubleshooting ({device_model}):")
-                        print("-> Try increasing timeout or checking network connectivity")
-                        print("-> Gateway may require different ARP command format")
-                    else:
-                        print(f"\nGeneral troubleshooting ({device_type}):")
-                        print("-> Check device connectivity and WebSocket support")
-                        print("-> Some devices may require SSH-based commands")
-
-                logging.warning("WebSocket ARP operation timed out")
-
-        except Exception as arp_error:
-            error_message = f"WebSocket ARP operation failed: {arp_error}"
-            print(f"! {error_message}")
-            logging.error(error_message)
-            logging.debug("EXIT: arp_device_websocket - error")
-
-        finally:
-            # Always cleanup WebSocket connection
-            try:
-                websocket_manager_local = locals().get("websocket_manager")
-                if websocket_manager_local is not None:
-                    websocket_manager_local.disconnect()
-                    print("-> WebSocket connection closed")
-            except Exception as cleanup_error:
-                logging.warning(f"WebSocket cleanup error: {cleanup_error}")
-
-            logging.debug("EXIT: arp_device_websocket")
-
-    @staticmethod
-    def service_ping_device():  # type: ignore[no-untyped-def]
-        """
-        Execute service ping command on SSR gateway devices via WebSocket.
-        Service ping allows ping packets to follow the same path as specific services.
-
-        Follows the documented Mist API pattern for service ping commands:
-        1. Subscribe to WebSocket channel
-        2. POST service ping command with service-specific parameters
-        3. Receive results via WebSocket stream with session-based demultiplexing
-
-        SECURITY: Uses authenticated WebSocket connection with session-based
-        command demultiplexing for concurrent command safety.
-        """
-        manager = ServicePingManager()  # type: ignore[no-untyped-call]
-        return manager.execute()
-
-
-class WebSocketCommands:
-    """
-    WebSocket Commands Class for Mist API device operations.
-
-    This class organizes all WebSocket-based device command functions following
-    the agents guide requirement that "All features, or helpers need to live
-    under the appropriately titled/named 'Class's for code clarity and organization."
-
-    All methods are static since they don't require instance state and can be
-    called directly from menu actions.
-
-    SECURITY: All methods use authenticated WebSocket connections with session-based
-    command demultiplexing for concurrent command safety.
-    """
-
-    @staticmethod
-    def show_mac_table():  # type: ignore[no-untyped-def]  # noqa: C901, PLR0912, PLR0915
-        """
-        Execute show MAC table command on a switch device via WebSocket.
-
-        MAC tables are a Layer 2 switching feature and are only meaningful on switches.
-        Routers/gateways operate at Layer 3 and typically don't maintain MAC tables.
-
-        Follows the documented Mist API pattern:
-        1. Connect to WebSocket
-        2. Subscribe to device command channel
-        3. Issue POST show_mac_table command
-        4. Await results via WebSocket stream
-
-        SECURITY: Uses authenticated WebSocket connection with session-based
-        command demultiplexing for concurrent command safety.
-        """
-        logging.info("Menu #5: Starting WebSocket show MAC table operation")
-        # Check for debug mode from command line arguments
-        debug_mode = "--debug" in sys.argv or "-d" in sys.argv
-
-        if debug_mode:
-            logging.getLogger().setLevel(logging.DEBUG)
-            print("[DEBUG] DEBUG MODE ENABLED")
-
-        logging.debug("ENTER: show_mac_table_websocket")
-
-        try:
-            # Interactive site selection
-            site_id = PromptUtils.select_site_id_from_csv()
-            if not site_id:
-                print("! No site selected. Operation cancelled.")
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Selected site_id = {site_id}")
-
-            # Get device selection - MAC table is a Layer 2 switching feature
-            print("-> MAC table is available on switches (Layer 2 devices)")
-            print("-> Routers/gateways operate at Layer 3 and typically don't maintain MAC tables")
-            print("-> APs forward wireless traffic but don't maintain traditional MAC tables")
-            device_id = PromptUtils.select_device_id_from_inventory(site_id, device_type="switch")
-            if not device_id:
-                print("! No switch device selected. MAC table command requires Layer 2 switching devices.")
-                print("! Only switches maintain MAC address learning tables for Ethernet forwarding.")
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Selected device_id = {device_id}")
-
-            print(f"\n-> Executing show MAC table on device {device_id}...")
-            print("-> Establishing WebSocket connection...")
-
-            # Initialize WebSocket manager
-            websocket_manager = WebSocketManager(apisession)
-
-            if debug_mode:
-                print("[DEBUG] WebSocketManager initialized")
-
-            # Connect to WebSocket
-            if not websocket_manager.connect():
-                print("! Failed to establish WebSocket connection")
-                return
-
-            if debug_mode:
-                print("[DEBUG] WebSocket connection established")
-
-            # Subscribe to device command channel
-            command_channel = f"/sites/{site_id}/devices/{device_id}/cmd"
-            if not websocket_manager.subscribe_to_channel(command_channel):
-                print("! Failed to subscribe to device command channel")
-                websocket_manager.disconnect()
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] Subscribed to channel: {command_channel}")
-
-            print("-> WebSocket connected and subscribed")
-
-            # Wait a moment for subscription to be established
-            time.sleep(1)
-
-            # Issue show MAC table command via REST API
-            mac_table_payload: dict[str, Any] = {}  # show_mac_table typically doesn't require additional parameters
-
-            print("-> Issuing show MAC table command...")
-            logging.debug(f"MAC table payload: {mac_table_payload}")
-
-            if debug_mode:
-                print(f"[DEBUG] MAC table payload = {mac_table_payload}")
-
-            # Get authentication details for direct HTTP request
-            mist_host = getattr(apisession, "host", None) or os.getenv("MIST_HOST")
-            mist_apitoken = getattr(apisession, "apitoken", None) or os.getenv("MIST_APITOKEN")
-
-            if not mist_host or not mist_apitoken:
-                print("! Mist host or API token not found in session or environment")
-                websocket_manager.disconnect()
-                return
-
-            if debug_mode:
-                print(f"[DEBUG] mist_host = {mist_host}")
-                print(f"[DEBUG] API token length = {len(mist_apitoken) if mist_apitoken else 0}")
-
-            # Make direct POST request to trigger show MAC table
-            mac_table_url = f"https://{mist_host}/api/v1/sites/{site_id}/devices/{device_id}/show_mac_table"
-            headers = {"Authorization": f"Token {mist_apitoken}", "Content-Type": "application/json"}
-
-            if debug_mode:
-                print(f"[DEBUG] POST URL = {mac_table_url}")
-                print("[DEBUG] Headers = {'Authorization': 'Token [REDACTED]', 'Content-Type': 'application/json'}")
-
-            mac_table_response = requests.post(mac_table_url, headers=headers, json=mac_table_payload, timeout=30)
-
-            if debug_mode:
-                print(f"[DEBUG] HTTP Response Status = {mac_table_response.status_code}")
-                print(f"[DEBUG] HTTP Response Body = {mac_table_response.text}")
-
-            if mac_table_response.status_code != 200:
-                print(f"! Failed to issue show MAC table command: {mac_table_response.status_code}")
-                print(f"! Response: {mac_table_response.text}")
-                websocket_manager.disconnect()
-                return
-
-            # Extract session ID from response
-            response_data = mac_table_response.json()
-            session_id = response_data.get("session")
-            if not session_id:
-                print("! No session ID returned from show MAC table command")
-                websocket_manager.disconnect()
-                return
-
-            print(f"-> Show MAC table command issued (session: {session_id[:8]}...)")
-            print("-> Waiting for MAC table results...")
-
-            if debug_mode:
-                print(f"[DEBUG] Full session ID = {session_id}")
-                print("[DEBUG] Starting to wait for WebSocket results...")
-
-            # Wait for MAC table results via WebSocket (longer timeout for potentially large tables)
-            mac_table_result = websocket_manager.wait_for_command_result(session_id, timeout_seconds=60)
-
-            if debug_mode:
-                print(f"[DEBUG] wait_for_command_result returned: {mac_table_result is not None}")
-                if mac_table_result:
-                    print(f"[DEBUG] Result keys: {list(mac_table_result.keys())}")
-
-            if mac_table_result:
-                print("\n" + "=" * 60)
-                print("MAC TABLE RESULTS:")
-                print("=" * 60)
-
-                # Display raw output (this is where MAC table results come according to documentation)
-                raw_output = mac_table_result.get("raw", "")
-                if raw_output:
-                    print("RAW OUTPUT:")
-                    print("-" * 40)
-                    print(raw_output)
-
-                # Display any other output fields that might be present
-                output_fields = mac_table_result.get("Output", "")
-                if output_fields and output_fields != raw_output:
-                    print("\nOTHER OUTPUT:")
-                    print("-" * 40)
-                    print(output_fields)
-
-                # Show all available fields for debugging
-                available_fields = [key for key in mac_table_result.keys() if key not in ["raw", "Output", "session"]]
-                if available_fields:
-                    print(f"\nOTHER AVAILABLE FIELDS: {available_fields}")
-                    for field in available_fields:
-                        field_value = mac_table_result.get(field)
-                        if field_value:
-                            print(f"{field}: {field_value}")
-
-                if not raw_output and not output_fields:
-                    print("No output data received")
-                    print(f"Available result keys: {list(mac_table_result.keys())}")
-
-                print("=" * 60)
-
-                # Log the successful operation
-                logging.info("WebSocket show MAC table completed successfully")
-
-            else:
-                print("! Timeout waiting for MAC table results")
-                print("! This may indicate:")
-                print("  - The device doesn't support MAC table commands (common for routers/Layer 3 devices)")
-                print("  - The device is busy or not responding")
-                print("  - Network connectivity issues")
-                print("! Note: MAC tables are primarily a Layer 2 (switch) feature")
-                logging.warning("WebSocket show MAC table operation timed out")
-
-                if debug_mode:
-                    print("[DEBUG] Checking WebSocket manager state...")
-                    print(f"[DEBUG] Connected = {websocket_manager.connected}")
-                    print(f"[DEBUG] Subscribed channels = {websocket_manager.subscribed_channels}")
-                    with websocket_manager.results_lock:
-                        print(f"[DEBUG] Pending results = {list(websocket_manager.command_results.keys())}")
-
-        except Exception as mac_table_error:
-            error_message = f"WebSocket show MAC table operation failed: {mac_table_error}"
-            print(f"! {error_message}")
-            logging.error(error_message)
-
-            if debug_mode:
-                print("[DEBUG] Exception details:")
-                import traceback
-
-                traceback.print_exc()
-
-            logging.debug("EXIT: show_mac_table_websocket - error")
-
-        finally:
-            # Always cleanup WebSocket connection
-            try:
-                websocket_manager_local = locals().get("websocket_manager")
-                if websocket_manager_local is not None:
-                    websocket_manager_local.disconnect()
-                    print("-> WebSocket connection closed")
-
-                    if debug_mode:
-                        print("[DEBUG] WebSocket cleanup completed")
-            except Exception as cleanup_error:
-                logging.warning(f"WebSocket cleanup error: {cleanup_error}")
-
-            logging.debug("EXIT: show_mac_table_websocket")
-
-    @staticmethod
-    def show_forwarding_table():  # type: ignore[no-untyped-def]
-        """
-        Execute show forwarding table command on a gateway/SSR device via WebSocket.
-
-        Displays Layer 3 routing table information for gateway/SSR devices.
-        This is a routing table diagnostic command specifically for devices that
-        perform Layer 3 forwarding functions.
-
-        SECURITY: Uses authenticated WebSocket connection with session-based
-        command demultiplexing for concurrent command safety.
-        """
-        logging.info("Menu #6: Starting WebSocket show forwarding table operation")
-        return RoutingUtils.execute_show_forwarding_table()  # type: ignore[no-untyped-call]
-
-    @staticmethod
-    def show_routing_table():  # type: ignore[no-untyped-def]
-        """
-        Execute show route command on switches, routers, and SSR devices via WebSocket.
-
-        Displays routing table information (RIB - Routing Information Base) for network devices.
-        This shows the routing protocol information maintained by routing protocols like BGP, OSPF,
-        static routes, etc. Different from forwarding table (FIB) which shows actual forwarding entries.
-
-        Supported on:
-        - Switches with Layer 3 capabilities
-        - SRX routers
-        - SSR gateways
-        - Other routing-capable devices
-
-        SECURITY: Uses authenticated WebSocket connection with session-based
-        command demultiplexing for concurrent command safety.
-        """
-        logging.info("Menu #7: Starting WebSocket show routing table operation")
-        return RoutingUtils.execute_show_routing_table()  # type: ignore[no-untyped-call]
-
-    @staticmethod
-    def show_ssr_routes():  # type: ignore[no-untyped-def]
-        """
-        Execute SSR/SRX routing table command using dedicated API function.
-
-        Uses the dedicated mistapi.api.v1.sites.devices.showSiteSsrAndSrxRoutes function
-        which provides structured routing table queries specifically optimized for
-        SSR and SRX devices with proper parameter validation and device-specific formatting.
-
-        This is the preferred method for SSR/SRX routing table queries as it provides:
-        - Structured parameter input (protocol, neighbor, prefix, vrf, node, route direction)
-        - Device-specific optimization for SSR/SRX platforms
-        - Proper BGP neighbor route analysis (received/advertised)
-        - VRF-aware routing table queries
-        - HA cluster node selection for multi-node deployments
-
-        Supported devices:
-        - SSR gateways (128T Session Smart Routers)
-        - SRX routers (Juniper SRX series)
-
-        SECURITY: Uses authenticated API session with proper parameter validation
-        and device capability checking for safe routing table operations.
-        """
-        logging.info("Menu #8: Starting SSR/SRX routing table query")
-        return RoutingUtils.execute_show_ssr_routes()  # type: ignore[no-untyped-call]
 
 
 class ServicePingManager:
@@ -30603,6 +28819,20 @@ class SiteInventoryHealthAnalyzer:
             print("! No sites found with offline infrastructure (all switches and gateways are online)")
 
 
+def _ws_cmd_deps() -> WebSocketCmdDeps:
+    """Create WebSocket command dependency context for the dispatch table."""
+    import mistapi.api.v1.sites.devices as _site_devices  # noqa: PLC0415
+
+    return WebSocketCmdDeps(
+        apisession=apisession,
+        select_site_fn=PromptUtils.select_site_id_from_csv,
+        select_device_fn=PromptUtils.select_device_id_from_inventory,
+        validate_target_fn=ValidationUtils.validate_ping_target,
+        list_devices_fn=_site_devices.listSiteDevices,
+        safe_input_fn=InputUtils.safe_input,
+    )
+
+
 menu_actions = {
     # ==============================
     # SYSTEM OPERATIONS
@@ -30623,17 +28853,20 @@ menu_actions = {
         "Export gateway management overlay IPs grouped by template association",
     ),
     # > WebSocket Device Commands
-    "5": (WebSocketCommands.show_mac_table, "Show MAC table on switch device via WebSocket (Layer 2 switching table)"),
+    "5": (
+        lambda: WebSocketCommands.show_mac_table(_ws_cmd_deps()),  # type: ignore[misc]
+        "Show MAC table on switch device via WebSocket (Layer 2 switching table)",
+    ),
     "6": (
-        WebSocketCommands.show_forwarding_table,
+        RoutingUtils.execute_show_forwarding_table,
         "Show forwarding table on gateway device via WebSocket (Layer 3 routing table)",
     ),
     "7": (
-        WebSocketCommands.show_routing_table,
+        RoutingUtils.execute_show_routing_table,
         "Show routing table on switches via WebSocket (Switch L3 routing - BGP/OSPF/Static)",
     ),
     "8": (
-        WebSocketCommands.show_ssr_routes,
+        RoutingUtils.execute_show_ssr_routes,
         "Show SSR/SRX routing table via dedicated API (128T/SRX gateways - Advanced BGP analysis)",
     ),
     # > Packet Capture Operations
@@ -30888,15 +29121,15 @@ menu_actions = {
         "Export Site Client Anomaly Events (client-specific anomaly detection: connectivity, roaming, throughput)",
     ),
     "87": (
-        WebSocketNetworkDiagCommands.ping_device,
+        lambda: WebSocketNetworkDiagCommands.ping_device(_ws_cmd_deps()),  # type: ignore[misc]
         "WebSocket Device Ping - Execute ping command on device via WebSocket stream (real-time output)",
     ),
     "88": (
-        WebSocketNetworkDiagCommands.arp_device,
+        lambda: WebSocketNetworkDiagCommands.arp_device(_ws_cmd_deps()),  # type: ignore[misc]
         "WebSocket Device ARP - Execute ARP command on device via WebSocket stream (real-time output)",
     ),
     "89": (
-        WebSocketNetworkDiagCommands.service_ping_device,
+        lambda: ServicePingManager().execute(),  # type: ignore[misc]
         "WebSocket Service Ping - Execute service-specific ping on SSR gateways via WebSocket stream (real-time output)",  # noqa: E501
     ),
     # ==============================
