+++ src/websocket/__init__.py
@@ -1,0 +1,13 @@
+"""WebSocket module for Mist API real-time communications."""
+
+from src.websocket.commands import WebSocketCommands
+from src.websocket.context import WebSocketCmdDeps
+from src.websocket.diag_commands import WebSocketNetworkDiagCommands
+from src.websocket.manager import WebSocketManager
+
+__all__ = [
+    "WebSocketCmdDeps",
+    "WebSocketCommands",
+    "WebSocketManager",
+    "WebSocketNetworkDiagCommands",
+]
